Final_demo/demo/src/main/backend/requirements.txt
flask>=2.0
flask-cors
ultralytics
pillow
numpy
reportlab
opencv-python
scipy
